Filter DocumentTypes; Bridge Translations; PersonDetailView; CSV Repair, DetailViewUpdate
Detail View Update:
Original Author shown for Translations
Actual Originals shown for translations of bridge translation
Online Filings shown only on the side with link in the new tab
Alphabetic order for filings
CSV repair
Link to PersonDetailView for Author and Translator

PersonDetailView:
so far with Alias: T/F and Organisation T/F (idk if needed)
Links to DocumentDetailView
In the future black column on the side with possibility to edit. So far nothing

Search:
Single Result are tagged as Translation/Original/Bridge (not sure if wanted)
When  string searched -> marked in the titles (easier to find wished document)

Filter by Document Type: Original/Translation/Bridge (must be seperately to pass it to another filters) Maybe some text above in the future? (Not sure if it is self explanatory for the user)
Filter by language
Choose type for sort (translation/original)