Final Bugfix
It turned out that the bug was due to a lack of testing. I only resubmitted the search form to test functionalities, but i did not attempt to enter the search view from the main page. This led to problems, as the code could not ask for data of a form that has never been submitted in the first place. Hopefully it works now